pyproject.toml
[project]
name = "attp-sdk"
version = "0.1.0"
description = "attp-sdk"
authors = [
    {name = "Zahcoder34",email = "Zakhriw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "ascender-framework (>=2.0.4,<3.0.0)",
    "attp-core (>=0.2.4,<0.3.0)",
    "msgpack (>=1.1.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
